+++ otserv/branches/revbattlesys/npc.h
@@ -111,7 +111,7 @@ public:
 
 	virtual bool isPushable() const { return true;};
 	
-	virtual unsigned long idRange(){ return 0x30000000;}
+	virtual uint32_t idRange(){ return 0x30000000;}
 	static AutoList<Npc> listNpc;
 	void removeList() {listNpc.removeList(getID());}
 	void addList() {listNpc.addList(this);}
