+++ otserv/branches/revbattlesys/weapons.cpp
@@ -459,7 +459,7 @@ bool Weapon::executeUseWeapon(Player* player, const LuaVariant& var) const
 
 	bool isSuccess = true;
 
-	long result;
+	int32_t result;
 	if(m_scriptInterface->callFunction(2, result) == false){
 		isSuccess = false;
 	}
